Refactor PubMedClient from monolithic client.rs into focused modules
Split the 2,920-line pubmed/client.rs god file into a client/ directory
module with 8 focused submodules, each owning a single E-utility API:

- client/mod.rs: Core struct, constructors, search, fetch, make_request (~400 lines)
- client/summary.rs: ESummary API (fetch_summaries, parse_esummary_response)
- client/history.rs: History server (EPost, fetch_from_history, search_all stream)
- client/einfo.rs: EInfo API (get_database_list, get_database_info)
- client/elink.rs: ELink API (get_related_articles, get_pmc_links, get_citations)
- client/citmatch.rs: ECitMatch API (match_citations)
- client/egquery.rs: EGQuery API (global_query)
- client/espell.rs: ESpell API (spell_check, spell_check_db)

All tests moved alongside their code. Public API unchanged - PubMedClient
methods are split across impl blocks in sibling modules using pub(crate)
accessors for internal fields.
